requirements.txt
pyyaml
selenium
pillow
